+++ CHANGELOG.rst
@@ -22,6 +22,8 @@ Changed
 
 - Check Raft failover availability on validate_config.
 
+- Forbid to enable Raft failover with ``ALL_RW`` replicasets.
+
 ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
 Added
 ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~